--fixed  table view toggle button act as another select method
